Rename all CSS modules to .module.scss.
