+++ app/components/ui/Banner/Banner.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './Banner.scss';
+import styles from './Banner.module.scss';
 /* eslint-disable max-len */
 
 export default function Banner() {
