+++ app/components/ui/Banner/Banner.module.scss
