+++ app/pages/Covid/Covid.module.scss
