+++ app/pages/Covid/components/CovidHeader/CovidHeader.module.scss
