+++ app/pages/Covid/components/FoodList/FoodList.module.scss
