+++ app/pages/Covid/components/FoodMap/FoodMap.module.scss
