+++ app/pages/HomePage/components/ResourceCard/ResourceCard.module.scss
