+++ app/pages/HomePage/components/ResourceList/ResourceList.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import ResourceCard from '../ResourceCard/ResourceCard';
-import styles from './ResourceList.scss';
+import styles from './ResourceList.module.scss';
 
 const ResourceList = ({ resources }) => (
   <ul className={styles.list}>
