+++ app/pages/HomePage/components/ResourceList/ResourceList.module.scss
