+++ app/pages/HomePage/components/SearchBar/SearchBar.module.scss
