+++ app/pages/HomePage/components/Section/Section.module.scss
