+++ app/pages/ServiceDiscoveryResults/OpenNowFilter.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connectRefinementList } from 'react-instantsearch/connectors';
 import { getCurrentDayTime } from '../../utils/index';
-import styles from './ServiceDiscoveryResults.scss';
+import styles from './ServiceDiscoveryResults.module.scss';
 
 
 /**
